flask_social_auth/requirements.txt
Flask
Flask-Dance
Flask-Login
Flask-SQLAlchemy
SQLAlchemy-Utils
python-dotenv
blinker
mysql-connector-python
Flask-Migrate
flask_bcrypt
Flask-WTF
email-validator